[WEB] refactor : 폴더명 변경
public 폴더를 utils로 변경했습니다.
